Desire Lottery,  30 minutes.
Players choose one from 1 to 16 numbers, then the lottery will calculate a random number of lottery numbers, if the player selected number is lottery figures, then the player win.
Return rate of up to 88%, is the highest of all types of lottery. The theoretical average prize money is 13.6 times the amount of the face amount.
From time to time 30 minutes, if the number of participants to reach the rated number or more, you can immediately lottery.
Always winning number, is a comprehensive player information, the ether square block information, as well as time information, calculated out. As the block and time information itself has a strong randomness, so to ensure the final winning numbers of strong randomness.